Фикс компонента формирования отчетов
